:wrench: moved all analytics to repository

+++ app/src/main/java/com/shub39/grit/habits/domain/HabitWithAnalytics.kt
@@ -15,5 +15,5 @@ data class HabitWithAnalytics(
     val weekDayFrequencyData: WeekDayFrequencyData,
     val currentStreak: Int,
     val bestStreak: Int,
-    val startedDaysAgo: Int
+    val startedDaysAgo: Long
 )
